Move void teleport to GameFixes
